+++ src/engine/components/numberGenerator/index.stories.jsx
@@ -12,6 +12,7 @@ import { Container } from 'pixi.js';
 import { storiesOf } from 'test/storybook-facades';
 import PixiWrapper from 'test/PixiWrapper';
 import { TEN, ONE } from 'constants/numbers';
+import blockGeneratorConfig, { BLOCK_DEFINITIONS } from 'engine/components/blockGeneratorConfig';
 
 import numberGenerator from './';
 
@@ -22,6 +23,7 @@ const SIZES = {
   negativeOne: 256,
   negativeTen: 512,
   elCaido: 128,
+  hitting: 64,
 };
 const MARGINS = {
   one: 10,
@@ -29,6 +31,7 @@ const MARGINS = {
   negativeOne: 30,
   negativeTen: 40,
   elCaido: 20,
+  hitting: 10,
 };
 const one = numberGenerator(ONE, POSITION, POSITION, SIZES.one, MARGINS.one).view;
 const ten = numberGenerator(TEN, POSITION, POSITION, SIZES.ten, MARGINS.ten).view;
@@ -51,13 +54,13 @@ storiesOf('engine.components.Number', module)
     />
   ))
   .add('happy negative one digit', () => {
-    const negativeOne = numberGenerator(-ONE, POSITION, POSITION, SIZES.negativeOne, MARGINS.negativeOne);
+    const number = numberGenerator(-ONE, POSITION, POSITION, SIZES.negativeOne, MARGINS.negativeOne);
 
-    negativeOne.beHappy('start');
+    number.beHappy('start');
 
     return (
       <PixiWrapper
-        component={negativeOne.view}
+        component={number.view}
         isContainer={false}
         height={SIZES.negativeOne + MARGINS.negativeOne + MARGINS.negativeOne}
         width={SIZES.negativeOne + MARGINS.negativeOne + MARGINS.negativeOne}
@@ -65,13 +68,13 @@ storiesOf('engine.components.Number', module)
     );
   })
   .add('sad negative two digits', () => {
-    const negativeTen = numberGenerator(-TEN, POSITION, POSITION, SIZES.negativeTen, MARGINS.negativeTen);
+    const number = numberGenerator(-TEN, POSITION, POSITION, SIZES.negativeTen, MARGINS.negativeTen);
 
-    negativeTen.beSad('start');
+    number.beSad('start');
 
     return (
       <PixiWrapper
-        component={negativeTen.view}
+        component={number.view}
         isContainer={false}
         height={SIZES.negativeTen + MARGINS.negativeTen + MARGINS.negativeTen}
         width={SIZES.negativeTen + MARGINS.negativeTen + MARGINS.negativeTen}
@@ -79,17 +82,17 @@ storiesOf('engine.components.Number', module)
     );
   })
   .add('El Caído, the _fallen one_', () => {
-    const negativeTen = numberGenerator(-TEN, POSITION, POSITION, SIZES.elCaido, MARGINS.elCaido);
+    const number = numberGenerator(-TEN, POSITION, POSITION, SIZES.elCaido, MARGINS.elCaido);
     const container = new Container();
 
-    container.addChild(negativeTen.view);
+    container.addChild(number.view);
 
     container.x = SIZES.elCaido / 1.25;
     container.y = SIZES.elCaido;
 
     (async () => {
-      await wait(2000);
-      await negativeTen.beTheFallenOne();
+      await wait(1000);
+      await number.beTheFallenOne();
     })();
 
     return (
@@ -98,6 +101,33 @@ storiesOf('engine.components.Number', module)
         isContainer={false}
         height={SIZES.elCaido * 3}
         width={SIZES.elCaido * 3}
+      />
+    );
+  })
+  .add('Hitting the Wall', () => {
+    const number = numberGenerator(-99, POSITION, POSITION, SIZES.hitting, MARGINS.hitting);
+    const block = blockGeneratorConfig(
+      BLOCK_DEFINITIONS.PATH,
+      SIZES.hitting,
+      MARGINS.hitting,
+    )(SIZES.hitting / 1.25, SIZES.hitting);
+
+    block.view.addChild(number.view);
+
+    (async () => {
+      await wait(1000);
+      await number.hitTheWall('top');
+      await number.hitTheWall('right');
+      await number.hitTheWall('bottom');
+      await number.hitTheWall('left');
+    })();
+
+    return (
+      <PixiWrapper
+        component={block.view}
+        isContainer={false}
+        height={SIZES.hitting * 3}
+        width={SIZES.hitting * 3}
       />
     );
   });