+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## HEAD
 
+- ⚙ [2017-05-31] `engine/containers/mazeEngineGenerator`: update to v4.2.0.
+  - ⚙ Animate the number on `MazePathError` error.
+- 🎨 [2017-05-31] `engine/components/blockGeneratorConfig`: update to v2.0.1.
+  - 🎨 Change `activePathBorders` param to optative. Add empty object as default value.
+- 🚀 [2017-05-31] `engine/containers/numberGenerator`: update to v4.2.0.
+  - 🚀 Add `hitTheWall` animation function.
 - ⚙ [2017-05-31] `engine/containers/numberGenerator`: update to v4.1.0.
   - ⚙ Update `hasEnteredToNumericLine` to add jump effect.
 - ⚙ [2017-05-31] `engine/containers/mazeEngineGenerator`: update to v4.1.0.